% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.arff_data}
\alias{print.arff_data}
\alias{summary.arff_data}
\title{Display functions}
\usage{
\method{print}{arff_data}(x, ...)

\method{summary}{arff_data}(object, max_attrs = 10, max_values = 5,
  ...)
}
\arguments{
\item{x}{A data.frame read from an ARFF file}

\item{...}{Extra parameters for the corresponding S3 method for class
\code{data.frame}}

\item{object}{A data.frame read from an ARFF file}
}
\description{
Display functions
}
