% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read.arff}
\alias{read.arff}
\title{Read an ARFF file}
\usage{
read.arff(file, stringsAsFactors = default.stringsAsFactors())
}
\arguments{
\item{file}{Name of the file to read the data from}

\item{stringsAsFactors}{logical: should categorical attributes be converted to factors?}
}
\value{
A `data.frame` with some attributes:
  - attributes: a named vector indicating the type of each variable
  - relation: the original "@relation" of the dataset
  - name: the extracted name (if the file is in MEKA format)
}
\description{
Reads a multilabel dataset from an ARFF file in Mulan or MEKA
and retrieves instances distinguishing attributes corresponding to labels
}
\examples{

library(yarr)
\dontrun{
yeast <- read.arff("yeast.arff")
}
}
