% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{attr.names}
\alias{attr.names}
\alias{attr.types}
\alias{relation}
\title{Dataset utilities}
\usage{
attr.names(x)

attr.types(x)

relation(x)
}
\arguments{
\item{x}{A dataset read using \code{read.arff}}
}
\description{
Some tools to access the metadata in ARFF files. \code{attr.names}
retrieves the names of the attributes, \code{attr.types} returns the ARFF type
of each variable, \code{relation} shows the name/relation specified in the
\code{@relation} section of the file.
}
