% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write.arff}
\alias{write.arff}
\title{Write a data.frame onto an ARFF file}
\usage{
write.arff(x, relation = NULL, types = NULL, file = "",
  sparse = FALSE, append = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}

\item{relation}{Name of the dataset}

\item{types}{A character vector indicating the type of each variable (optional)}

\item{file}{Name of the file to read the data from}

\item{sparse}{Logical: write in sparse format?}

\item{append}{Logical: append to an existing file?}

\item{...}{Extra parameters for internal functions}
}
\value{
A `data.frame` with some attributes:
  - attributes: a named vector indicating the type of each variable
  - relation: the original "@relation" of the dataset
  - name: the extracted name (if the file is in MEKA format)
}
\description{
Takes a data frame and records it in ARFF (Attribute-Relation File Format).
}
\examples{

library(yarr)
\dontrun{
write.arff(iris, "iris", file = "iris.arff")
}
}
